+++ payments-core/src/main/java/com/stripe/android/StripeKtx.kt
@@ -8,7 +8,7 @@ import com.stripe.android.core.exception.InvalidRequestException
 import com.stripe.android.core.exception.StripeException
 import com.stripe.android.core.model.StripeModel
 import com.stripe.android.core.networking.ApiRequest
-import com.stripe.android.exception.AuthenticationException
+import com.stripe.android.core.exception.AuthenticationException
 import com.stripe.android.exception.CardException
 import com.stripe.android.model.AccountParams
 import com.stripe.android.model.BankAccountTokenParams
