+++ stripe-core/src/main/java/com/stripe/android/core/exception/MaxRetryReachedException.kt
@@ -1,0 +1,10 @@
+package com.stripe.android.core.exception
+
+import androidx.annotation.RestrictTo
+
+/**
+ * An [Exception] that represents max retry is reached when making a request.
+ */
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+class MaxRetryReachedException(message: String? = null) :
+    StripeException(message = message)
